añadir organización carpetas en assets -> ASCII
