--- src/main/java/org/f5games/aiden_game/ascii/GameOver.java
@@ -1,21 +1,0 @@
-package org.f5games.aiden_game.ascii;
-
-public class GameOver {
-    public static void main(String[] args) {
-        System.out.println();
-        System.out
-                .println("- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -");
-        System.out.print("  ________    _____      _____   ___________ ________   ____   _________________________ \n" + //
-                " /  _____/   /  _  \\    /     \\  \\_   _____/ \\_____  \\  \\   \\ /   /\\_   _____/\\______   \\\n"
-                + //
-                "/   \\  ___  /  /_\\  \\  /  \\ /  \\  |    __)_   /   |   \\  \\   Y   /  |    __)_  |       _/\n" + //
-                "\\    \\_\\  \\/    |    \\/    Y    \\ |        \\ /    |    \\  \\     /   |        \\ |    |   \\\n"
-                + //
-                " \\______  /\\____|__  /\\____|__  //_______  / \\_______  /   \\___/   /_______  / |____|_  /\n" + //
-                "        \\/         \\/         \\/         \\/          \\/                    \\/         \\/ ");
-        System.out.println();
-        System.out
-                .println("- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -");
-        System.out.println();
-    }
-}
