+++ src/main/java/org/f5games/aiden_game/assets/ascii/Skeletons.java
@@ -1,4 +1,4 @@
-package org.f5games.aiden_game.ascii;
+package org.f5games.aiden_game.assets.ascii;
 
 public class Skeletons {
     public static void main(String[] args) {
