+++ src/main/java/org/f5games/aiden_game/assets/text/Intro.java
@@ -1,0 +1,14 @@
+package org.f5games.aiden_game.assets.text;
+
+public class Intro {
+    public static void main(String[] args) {
+        System.out.println("- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -");
+        System.out.println("- -        ¡Bienvenido al juego de La Noche De Los Espíritus!           - -");
+        System.out.println("- -                                                                     - -");
+        System.out.println("- - Intrucciones: 3 niveles completos: Bosque, Cementerio y Castillo    - -");
+        System.out.println("- - Aiden comienza con MOCHILA VACÍA [ _ , _ , _ ]                      - -");
+        System.out.println("- - Habrá un jefe por cada nivel hasta llegar al jefe final MORTIS      - -");
+        System.out.println("- -                                                                     - -");
+        System.out.println("- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -");
+    }
+}
