+++ src/main/java/org/f5games/aiden_game/assets/text/Level1.java
@@ -1,0 +1,9 @@
+package org.f5games.aiden_game.assets.text;
+
+import org.f5games.aiden_game.assets.ascii.Forest;
+
+public class Level1 {
+    public static void main(String[] args) {
+        Forest.main(args);
+    }
+}
