+++ src/main/java/org/f5games/aiden_game/assets/text/Level2.java
@@ -1,0 +1,9 @@
+package org.f5games.aiden_game.assets.text;
+
+import org.f5games.aiden_game.assets.ascii.Graveyard;
+
+public class Level2 {
+    public static void main(String[] args) {
+        Graveyard.main(args);
+    }
+}
