+++ src/main/java/org/f5games/aiden_game/assets/text/Level3.java
@@ -1,0 +1,9 @@
+package org.f5games.aiden_game.assets.text;
+
+import org.f5games.aiden_game.assets.ascii.Castle;
+
+public class Level3 {
+    public static void main(String[] args) {
+        Castle.main(args);
+    }
+}
